WIP: don't pick up macOS as +macOS, since that will override any attempt to change style
The style engine uses file selectors to get the correct style controls. This is why
it falls back to selecting the default style if a specific style does not contain
a control. The problem is that QQC2 has overriden the file selector to avoid the
need to prepend '+' in front of a folder. The side effect is that it will pick
up 'macOS' before any other style, since it matches the name of the os.

So we need to change this. Since we still wan't application to be able to
create styles that makes use of file selectors internally, we should
probably change / divide how we search for style controls in QQC2.